requirements.txt
requests
python-telegram-bot
googletrans==3.1.0a0